% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/acoustic_array.R
\name{acoustic_array}
\alias{acoustic_array}
\title{Create objects of class `acoustic_array`}
\usage{
acoustic_array(station_names, station_positions)
}
\arguments{
\item{station_names}{vector containing the station names (or receiver ids) in
the array}

\item{station_positions}{numeric matrix, data.frame or SpatialPoints object
  with the coordinates (Longitude/Latitude) of the stations. Each row is a
  station and there must be only one coordinate per station}
}
\value{
An object of the class `acoustic array`. It is a special case of a
  data.frame with columns 'station', 'lon' and 'lat'
}
\description{
Creates an object of the class `acoustic_array`. It ensures
  compatibility with other functions in the package
}
\examples{
plot_crayons()
}

