% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/detections.R
\name{detections}
\alias{detections}
\title{Create objects of class `detections`}
\usage{
detections(det_times, det_stations, det_tag = NULL, det_sensor = NULL, data)
}
\arguments{
\item{det_times}{vector containing the times of the detections. If the vector
is not an Date-Time (POSIXct) object, it would be coerced to POSIXct using
\code{\link[readr]{parse_datetime}}. Alternatively, if the argument `data`
is provided it can be the name (quoted or unquoted) of the corresponding
column in data}

\item{det_stations}{a vector with the names of the receiver id/stations name
where the tag was detected. Alternatively, if the argument `data` is
provided it can be the name (quoted or unquoted) of the corresponding
column in data}

\item{det_tag}{an (optional) vector containing the transmitter or animal ids.
Alternatively, if the argument `data` is provided it can be the name
(quoted or unquoted) of the corresponding column in data}

\item{det_sensor}{an (optional) vector containing sensor information.
Alternatively, if the argument `data` is provided it can be the name
(quoted or unquoted) of the corresponding column in data}

\item{an}{(optional) data frame containing the detection information. If
  provided, det_stations, det_times, det_tag and det_sensor must correspond
  to (quoted or unquoted) column names}
}
\value{
A `detections` object. A special case of a `data.frame` with standard
  columns: `station`, `time`, `tag`, `sensor`
}
\description{
This function creates an object of the class `detections` to be
  passed as argument of other functions that require information about the
  location of the stations. It ensures that the formating and structure of
  the detection data is consistent and makes it compatible with all other
  functions in this package
}
\examples{
plot_crayons()
}

