% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_smooth.R
\name{get_ssm_smooth}
\alias{get_ssm_smooth}
\title{Title}
\usage{
get_ssm_smooth(ssm_results, state_space_grid)
}
\arguments{
\item{ssm_results}{}

\item{state_space_grid}{}
}
\value{
jj
}
\description{
Title
}

