% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/range_test.R
\name{range_test}
\alias{range_test}
\title{Create objects of class `range_test`}
\usage{
range_test(x)
}
\arguments{
\item{x}{logistic relationship of the detection probability as a function of
  distance from a receiver. It can be specified as 1) a named list with the
  fields P0 (detection probability at a distance 0), D50 (distance
  corresponding to a detection probability of 50\%), and D95 (distance
  corresponding to a detection probability of 5\%). OR 2) a numeric vector
  with three elements corresponding to P0, D50 and D95. OR 3) a
  \code{\link{glm}} model of with a binomial family}
}
\value{
j
}
\description{
jhjh
}

