% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ssm.R
\name{ssm}
\alias{ssm}
\title{ssm}
\usage{
ssm(detections, acoustic_array, range_test, sentinel, delta_t,
  lon_grid = list(from = NA, to = NA, extra = 0.2), lat_grid = list(from =
  NA, to = NA, extra = 0.2), cell_width = NULL, n_cells = 2500,
  start_time = NULL, end_time = NULL, safe = FALSE, verbose = FALSE,
  admb_errors = c("stop", "warn", "ignore"))
}
\value{
jj
}
\description{
ssm
}
\examples{
#
}

